Refactor code to support use as a library
* Move the options code into pkg because we want to be able to use some of the structs
  in other packages

* The main use case is creating hmac signatures for use in test cases of githubapps
