+++ pkg/handlers.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"crypto/tls"
